Bgg data fetcher
Currently just grabs a rating for a board game if you supply the game id. Should be easy to expand if I want to grab any other information from bgg API. I am curious if theres an easier way to get the id for the game itself.

When an admin makes a post for a board game, they will have to supply the game id so it's in the Post table. There might be a way to use the BGG API to search for a board game when making a post that would automatically fill out several columns for the table, most importantly the id number.